backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
openai==1.40.1
youtube-transcript-api==0.6.2
reportlab==4.0.9
python-pptx==0.6.23
